refactor: move show_error_dialog to gui/util.py
Extract the duplicate show_error_dialog implementation from
primer_dimer_view.py and result_view.py into a single shared helper
function inside gui/util.py.
